Refactor routes functionality into separate routes.go file
- Extract all routing table functions from main.go to routes.go
- Remove unused os/exec import from main.go after refactoring
- Maintain same functionality while following project conventions
- Functions moved: checkRoutes, getIPv4Routes, getIPv6Routes, formatRoute, formatRouteLinux
